requirements.txt
pandas
flask
numpy
requests
textblob
gunicorn
python-dotenv
